Remove all fancy plugins from vimrc
I'd like to try out VSCode for development and that will be easier to do
if I make nvim less functional:

- Remove all LSP support.
- Remove rust-analyzer
- Remove neovim provider plugins
- Remove editorconfig plugin
- Remove code completion plugins
- Remove nvim telescope

This change also backs out the changes to use nvim-telescope and
replaces it with the fzf plugin. nvim-telescope compiles a one-off
binary which is not going to play well with Santa at work.